requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dateutil==2.9.0.post0
